+++ services/madoc-ts/src/frontend/shared/atoms/Carousel.tsx
@@ -1,20 +1,44 @@
 import * as React from 'react';
 import styled from 'styled-components';
 import { Button } from '../navigation/Button';
-import { Chevron } from '../icons/Chevron';
+import { ChevronLeft, ChevronRight } from '../icons/ChevronIcon';
 
 const CarouselOuterWrapper = styled.div`
   width: 90%;
+  max-width: 1100px;
 `;
 
 const CarouselWrapper = styled.div`
   display: flex;
   justify-content: space-between;
   align-items: center;
+`;
 
-  svg {
-    height: 10px;
+const CarouselControl = styled.button<{
+  $color?: string;
+}>`
+  border: none;
+  background-color: transparent;
+  color:  ${props => (props.$color ? props.$color : '#3579f6')};
+  
+  display: flex;
+  font-size: 1em;
+  padding: 1em;
+  
+  &:hover {
+    color: #333333;
+    cursor: pointer;
+    
+    svg {
+      fill: #333333;
+    }
   }
+  svg {
+    fill: ${props => (props.$color ? props.$color : '#3579f6')};
+    margin: 0 0.5em;
+    position: relative;
+    top: 0.15em;
+  },
 `;
 
 const CarouselSlides = styled.div`
@@ -22,11 +46,13 @@ const CarouselSlides = styled.div`
   width: 100%;
   display: flex;
   overflow: hidden;
+  justify-content: center;
+
 `;
 const CarouselSlide = styled.div`
   opacity: 0;
   width: 0;
-  transition: opacity 0.8s ease, transform 0.8s ease;
+  transition: opacity 0.5s ease, transform 0.8s ease;
   visibility: hidden;
 
   > div {
@@ -48,19 +74,23 @@ const CarouselSlide = styled.div`
   }
 `;
 
-const Indicator = styled.button`
+const Indicator = styled.button<{
+  $color?: string;
+}>`
   height: 12px;
   width: 64px;
   background-color: transparent;
-  border: 1px solid #3579f6;
+  border: 1px solid;
+  border-color: ${props => (props.$color ? props.$color : '#3579f6')};
   align-self: end;
   margin: 1em;
 
   :hover {
+    cursor: pointer;
     background-color: rgba(59, 59, 93, 0.7);
   }
   &[data-is-active='true'] {
-    background-color: #3579f6;
+    background-color: ${props => (props.$color ? props.$color : '#3579f6')};
 
     :hover {
       background-color: rgba(59, 59, 93, 0.7);
@@ -74,9 +104,10 @@ const IndicatorWrapper = styled.div`
 
 interface CarouselProps {
   children: React.ReactNode;
+  controlColor?: string;
 }
 
-export const Carousel = ({ children }: CarouselProps) => {
+export const Carousel = ({ children, controlColor }: CarouselProps) => {
   const [currentSlide, setCurrentSlide] = React.useState(0);
   const slides = React.Children.toArray(children);
 
@@ -97,29 +128,30 @@ export const Carousel = ({ children }: CarouselProps) => {
   return (
     <CarouselOuterWrapper>
       <CarouselWrapper>
-        <Button
-          $link={true}
+        <CarouselControl
+          $color={controlColor}
           onClick={() => {
             setCurrentSlide(prev);
           }}
         >
-          <Chevron style={{ transform: 'scaleX(-1)' }} />
+          <ChevronLeft />
           Previous
-        </Button>
+        </CarouselControl>
         <CarouselSlides>{activeSlide}</CarouselSlides>
-        <Button
-          $link={true}
+        <CarouselControl
+          $color={controlColor}
           onClick={() => {
             setCurrentSlide(next);
           }}
         >
           Next
-          <Chevron />
-        </Button>
+          <ChevronRight />
+        </CarouselControl>
       </CarouselWrapper>
       <IndicatorWrapper>
         {slides?.map((slide, index) => (
           <Indicator
+            $color={controlColor}
             data-is-active={currentSlide === index}
             key={index}
             onClick={() => {